Fix artifacting with badge font size override
On Chrome on Linux, the button text aligns to the bottom when
line-height is specified, and on Firefox on Linux, the input text gets
chopped off a bit on the top and bottom when an explicit height is
specified. Rather than use an explicit height, just adjust the font size
and padding slightly to achieve the same effect.